+++ src/index.ts
@@ -24,6 +24,7 @@ export {
 export {
   Conversation,
   ConversationState,
+  SendMessageObject,
   StateObject,
 } from './models/conversation';